Correction on TIFFTAG_SUBIFD byte size

+++ libtiff/libtiff_ctypes.py
@@ -375,7 +375,7 @@ tifftags = {
     TIFFTAG_TILEOFFSETS: (
         ctypes.POINTER(ctypes.c_uint32), lambda _d: _d.contents),
 
-    TIFFTAG_SUBIFD: (ctypes.c_uint32 * 0, lambda d:d[1][:d[0].value]),  # uint16*,uint32**  count & IFD arrays
+    TIFFTAG_SUBIFD: (ctypes.c_uint64 * 0, lambda d:d[1][:d[0].value]),  # uint16*,uint32**  count & IFD arrays
     TIFFTAG_BITSPERSAMPLE: (ctypes.c_uint16, lambda _d: _d.value),
     TIFFTAG_CLEANFAXDATA: (ctypes.c_uint16, lambda _d: _d.value),
     TIFFTAG_COMPRESSION: (ctypes.c_uint16, lambda _d: _d.value),
